+++ js/functions.js
@@ -1,21 +1,14 @@
 const findStringLength = (string, minLength) => String(string).length >= minLength;
 
-findStringLength('house', 5);
-
 const isPalindrome = (string) => {
   string = String(string).toLowerCase().replaceAll(' ', '');
   return string === string.split('').reverse().join('');
 };
 
-isPalindrome('Лёша на полке клопа нашёл ');
-
 const findNumber = (string) => {
   const onlyNumbers = String(string).replace(/\D/g, '');
   return parseInt(onlyNumbers, 10);
 };
-
-findNumber('-2023 year 0.3');
-findNumber('а я томат');
 
 const createString = (string, length, symbols) => {
   while(string.length < length) {
@@ -25,4 +18,4 @@ const createString = (string, length, symbols) => {
   return string;
 };
 
-createString('house', 10, 's');
+export {findStringLength, isPalindrome, findNumber, createString};
